[FE] fix : fetchFoodItems 수정

+++ front/D101/app/src/main/java/com/ssafy/d101/model/FoodInfo.kt
@@ -8,7 +8,7 @@ data class FoodInfo(
     val dbGroup: String,
     val manufacturer: String,
     val portionSize: Int,
-    val totalSize: Int,
+    val totalSize: Double,
     val unit: String,
     val calorie: Int,
     val carbohydrate: Double,
